requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
Django==3.2.4
django-classy-tags==2.0.0
django-cms==3.8.0
django-formtools==2.3
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-snippet==3.0.0
djangorestframework==3.12.4
idna==2.10
importlib-metadata==4.6.0
jira==3.0.1
keyring==23.0.1
oauthlib==3.1.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.6
zipp==3.4.1
